Basic interactable functionality
Created a abstract Interactable class with two functions: Interact and baseInteract and global string promptMessage. This abstract class will be used by subclasses. Keypad class is an example.
PlayerInteract class is used for the ray functionality of the camera. if there is any object 3f in front of the player, the player will detect it. Right now, if the player gets close to the keypad, it gives a message in the log.