Created custom interaction events
To make it a lot easier to create interaction event scripts for certain gameObjects. Here is what each class does:
The InteractionEvent script defines the Unity event that can be triggered on interaction.
The Interactable script serves as the base class for all objects that can be interacted with. It checks if events should be used, and if so, it triggers the event and calls a customizable Interact() method.
The EditorInteractable class is a custom editor that modifies how the Interactable objects (and their derived classes) are displayed in the Unity inspector:

    For objects of type EventOnlyInteractable, it forces the use of Unity events and ensures the InteractionEvent component is added.
    For other Interactable objects, it dynamically adds or removes the InteractionEvent component based on the useEvents flag.