Added an assault Rifle for the player
Added an assault rifle for the player. Changed gun settings to make it more automatic with a lot more bullets. Added a couple of animations; reload animation and recoil animation were adds similar to the pistol but I also added a reload animation. Muzzle flash, bullets and bullet holes also work  for this gun