Added pickup weapon and switching functionality
Player can now interact with a weapon in the world space and make the game object weapon attach to the player. The player can also switch between two different weapon slots available in the player gameObject.
The gun objects in the world space are highlighted. I used a free highlighter in unity asset store.