fixed some enemy animation bugs
When the enemy loses sight of the player, it would take about 8 seconds to go from attack state to search state, i chnaged it to 1 second, so the enemy would instantly look for the player when it loses sight of them.
I also fixed the search animation as i found a bug. When the enemy would search for the player, it would walk in a spot when it stopped. I made it so that when the enemy stopped, it would change to the idle animation and when it continued to search, it would revert back to the search animation.