Added bloodSpray particles
Downloaded and imported bloodSpray particles, so enemy shoots blood when getting shot.
Added an GameObjectDestroy script which just destroys a specific gameObject after whatever seconds is prescribed immediately after it has been instantiated (using it for the blood spray object).
Changed PlayerBullet, so that if an enemy is dead, the player cant still shoot at it causing effects like the blood splatter or animations.
Also added back the interaction range, so the player cant interact with certain objects halfway across the map.
Finally I fixed a bug where the player was able to equip the last gun it had replaced even though the player wasn't looking or interacting with the gun.