Imported new level assets
Imported new level space assets. This includes walls, floors, etc...
Added some quick room areas to see if the assets worked. That includes two rooms and a corridor.
Added the player prefab and managers. Set everything needed such as the UI and stuff. This was important to be able to move the player around the new level